+++ src/js/controllers.js
@@ -549,7 +549,7 @@ $scope.updateABV();
   };
 
   $scope.loadConfig = function(){
-    var config = [];
+    let config = [];
     if(!$scope.pkg){
       config.push(BrewService.pkg().then(function(response){
           $scope.pkg = response;
@@ -990,8 +990,7 @@ $scope.updateABV();
     $scope.resetError(kettle);
   };
 
-  function downloadSketch(actions, sketch, name){
-    console.log('name',name)
+  function downloadSketch(name, actions, headers, sketch){
     // tp link connection
     let tplink_connection_string = BrewService.tplink().connection();
     // influx db connection
@@ -1004,17 +1003,28 @@ $scope.updateABV();
       connection_string += `u=${$scope.settings.influxdb.user}&p=${$scope.settings.influxdb.pass}&`
     // add db
     connection_string += 'db='+($scope.settings.influxdb.db || 'session-'+moment().format('YYYY-MM-DD'));
-    let header = '/* Sketch Auto Generated from http://monitor.brewbench.co on '+moment().format('YYYY-MM-DD HH:MM:SS')+' for '+name+'*/\n';
+    let autogen = '/* Sketch Auto Generated from http://monitor.brewbench.co on '+moment().format('YYYY-MM-DD HH:MM:SS')+' for '+name+'*/\n';
+
     $http.get('assets/arduino/'+sketch+'/'+sketch+'.ino')
       .then(response => {
         // replace variables
-        response.data = header+response.data
+        response.data = autogen+response.data
           .replace('// [actions]', actions.length ? actions.join('\n') : '')
+          .replace('// [headers]', headers.length ? headers.join('\n') : '')
           .replace('[TPLINK_CONNECTION]', tplink_connection_string)
           .replace('[SLACK_CONNECTION]', $scope.settings.notifications.slack)
           .replace('[FREQUENCY_SECONDS]', $scope.settings.sketches.frequency ? parseInt($scope.settings.sketches.frequency,10) : 60);
         if( sketch.indexOf('InfluxDB') !== -1){
-          response.data = response.data.replace('[INFLUXDB_CONNECTION]', connection_string)
+          response.data = response.data.replace('[INFLUXDB_CONNECTION]', connection_string);
+        }
+        if(headers.indexOf('#include <dht.h>') !== -1){
+          response.data = response.data.replace(/\/\/ DHT /g, '');
+        }
+        if(headers.indexOf('#include <cactus_io_DS18B20.h>') !== -1){
+          response.data = response.data.replace(/\/\/ DS18B20 /g, '');
+        }
+        if(actions.length){
+          response.data = response.data.replace(/\/\/ triggers /g, '');
         }
         let streamSketch = document.createElement('a');
         streamSketch.setAttribute('download', sketch+'-'+name+'.ino');
@@ -1029,62 +1039,91 @@ $scope.updateABV();
   $scope.downloadAutoSketch = function(){
     let sketches = [];
     let actions = [];
-    let lastArduino = '';
-    _.each(_.orderBy($scope.kettles, 'arduino.url', 'asc'), (kettle, i) => {
-      lastArduino = kettle.arduino.url.replace(/[^a-zA-Z0-9-.]/g, "");
+    let headers = [];
+    let downloads = [];
+    let arduinoName = '';
+    _.each($scope.kettles, (kettle, i) => {
       // reset the actions
       if((kettle.heater && kettle.heater.sketch) ||
         (kettle.cooler && kettle.cooler.sketch) ||
         kettle.notify.dweet
       ){
-        if( sketches.indexOf(lastArduino) === -1 ) sketches.push(lastArduino);
-        // download previous sketch
-        if( sketches.length > 1 && lastArduino != sketches[sketches.length-2]){
-          downloadSketch(actions, 'BrewBenchAutoYun', sketches[sketches.length-2]);
-          // reset actions
-          actions = [];
+        arduinoName = kettle.arduino.url.replace(/[^a-zA-Z0-9-.]/g, "");
+        let currentSketch = _.find(sketches,{name:arduinoName});
+        if(!currentSketch){
+          sketches.push({
+            name: arduinoName,
+            actions: [],
+            headers: []
+          });
+          currentSketch = _.find(sketches,{name:arduinoName});
         }
         let target = ($scope.settings.unit=='F') ? $filter('toCelsius')(kettle.temp.target) : kettle.temp.target;
         let adjust = ($scope.settings.unit=='F' && kettle.temp.adjust != 0) ? Math.round(kettle.temp.adjust*0.555) : kettle.temp.adjust;
-        actions.push('temp = autoCommand("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'","'+kettle.temp.pin+'","'+kettle.temp.type+'",'+adjust+');');
+        if(kettle.temp.type.indexOf('DHT') !== -1){
+          currentSketch.headers.push('// https://www.brewbench.co/libs/DHTLib.zip');
+          currentSketch.headers.push('#include <dht.h>');
+        }
+        else if(kettle.temp.type.indexOf('DS18B20') !== -1){
+          currentSketch.headers.push('// https://www.brewbench.co/libs/cactus_io_DS18B20.zip');
+          currentSketch.headers.push('#include <cactus_io_DS18B20.h>');
+        }
+        currentSketch.actions.push('temp = autoCommand("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'","'+kettle.temp.pin+'","'+kettle.temp.type+'",'+adjust+');');
         //look for triggers
         if(kettle.heater && kettle.heater.sketch)
-          actions.push('trigger("heat","'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'","'+kettle.heater.pin+'",temp,'+target+','+kettle.temp.diff+','+!!kettle.notify.slack+');');
+          currentSketch.actions.push('trigger("heat","'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'","'+kettle.heater.pin+'",temp,'+target+','+kettle.temp.diff+','+!!kettle.notify.slack+');');
         if(kettle.cooler && kettle.cooler.sketch)
-          actions.push('trigger("cool","'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'","'+kettle.cooler.pin+'",temp,'+target+','+kettle.temp.diff+','+!!kettle.notify.slack+');');
+          currentSketch.actions.push('trigger("cool","'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'","'+kettle.cooler.pin+'",temp,'+target+','+kettle.temp.diff+','+!!kettle.notify.slack+');');
         if(kettle.notify.dweet)
-          actions.push('dweetAutoCommand("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'","'+$scope.settings.recipe.brewer.name+'","'+$scope.settings.recipe.name+'",temp);');
+          currentSketch.actions.push('dweetAutoCommand("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'","'+$scope.settings.recipe.brewer.name+'","'+$scope.settings.recipe.name+'",temp);');
       }
     });
-    downloadSketch(actions, 'BrewBenchAutoYun', sketches[sketches.length-1]);
+    _.each(sketches, (sketch, i) => {
+      downloadSketch(sketch.name, sketch.actions, sketch.headers, 'BrewBenchAutoYun');
+    });
   };
 
   $scope.downloadInfluxDBSketch = function(){
     if(!$scope.settings.influxdb.url) return;
     let sketches = [];
     let actions = [];
-    let lastArduino = '';
-    _.each(_.orderBy($scope.kettles, 'arduino.url', 'asc'), (kettle, i) => {
-      lastArduino = kettle.arduino.url.replace(/[^a-zA-Z0-9-.]/g, "");
-      if( sketches.indexOf(lastArduino) === -1 ) sketches.push(lastArduino);
-      // download previous sketch
-      if( sketches.length > 1 && lastArduino != sketches[sketches.length-2]){
-        downloadSketch(actions, 'BrewBenchInfluxDBYun', sketches[sketches.length-2]);
-        // reset actions
-        actions = [];
+    let headers = [];
+    let downloads = [];
+    let arduinoName = '';
+
+    _.each($scope.kettles, (kettle, i) => {
+      arduinoName = kettle.arduino.url.replace(/[^a-zA-Z0-9-.]/g, "");
+      let currentSketch = _.find(sketches,{name:arduinoName});
+      if(!currentSketch){
+        sketches.push({
+          name: arduinoName,
+          actions: [],
+          headers: []
+        });
+        currentSketch = _.find(sketches,{name:arduinoName});
       }
       let target = ($scope.settings.unit=='F') ? $filter('toCelsius')(kettle.temp.target) : kettle.temp.target;
       let adjust = ($scope.settings.unit=='F' && kettle.temp.adjust != 0) ? Math.round(kettle.temp.adjust*0.555) : kettle.temp.adjust;
-      actions.push('temp = influxDBCommand(F("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'"),F("'+kettle.temp.pin+'"),F("'+kettle.temp.type+'"),'+adjust+');');
+      if(kettle.temp.type.indexOf('DHT') !== -1){
+        currentSketch.headers.push('// https://www.brewbench.co/libs/DHTLib.zip');
+        currentSketch.headers.push('#include <dht.h>')
+      }
+      else if(kettle.temp.type.indexOf('DS18B20') !== -1){
+        currentSketch.headers.push('// https://www.brewbench.co/libs/cactus_io_DS18B20.zip');
+        currentSketch.headers.push('#include <cactus_io_DS18B20.h>')
+      }
+      currentSketch.actions.push('temp = influxDBCommand(F("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'"),F("'+kettle.temp.pin+'"),F("'+kettle.temp.type+'"),'+adjust+');');
       //look for triggers
       if(kettle.heater && kettle.heater.sketch)
-        actions.push('trigger(F("heat"),F("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'"),F("'+kettle.heater.pin+'"),temp,'+target+','+kettle.temp.diff+','+!!kettle.notify.slack+');');
+        currentSketch.actions.push('trigger(F("heat"),F("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'"),F("'+kettle.heater.pin+'"),temp,'+target+','+kettle.temp.diff+','+!!kettle.notify.slack+');');
       if(kettle.cooler && kettle.cooler.sketch)
-        actions.push('trigger(F("cool"),F("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'"),F("'+kettle.cooler.pin+'"),temp,'+target+','+kettle.temp.diff+','+!!kettle.notify.slack+');');
+        currentSketch.actions.push('trigger(F("cool"),F("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'"),F("'+kettle.cooler.pin+'"),temp,'+target+','+kettle.temp.diff+','+!!kettle.notify.slack+');');
       if(kettle.notify.dweet)
-        actions.push('dweetAutoCommand(F("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'"),F("'+$scope.settings.recipe.brewer.name+'"),F("'+$scope.settings.recipe.name+'"),temp);');
+        currentSketch.actions.push('dweetAutoCommand(F("'+kettle.key.replace(/[^a-zA-Z0-9-.]/g, "")+'"),F("'+$scope.settings.recipe.brewer.name+'"),F("'+$scope.settings.recipe.name+'"),temp);');
     });
-    downloadSketch(actions, 'BrewBenchInfluxDBYun', sketches[sketches.length-1]);
+    _.each(sketches, (sketch, i) => {
+      downloadSketch(sketch.name, sketch.actions, sketch.headers, 'BrewBenchInfluxDBYun');
+    });
   };
 
   $scope.getIPAddress = function(){